+++ ErsatzTV.Application/Subtitles/Queries/GetSubtitlePathByIdHandler.cs
@@ -1,5 +1,7 @@
 using Dapper;
 using ErsatzTV.Core;
+using ErsatzTV.Core.Domain;
+using ErsatzTV.Core.Jellyfin;
 using ErsatzTV.Infrastructure.Data;
 using ErsatzTV.Infrastructure.Extensions;
 using Microsoft.EntityFrameworkCore;
@@ -20,11 +22,32 @@ public class GetSubtitlePathByIdHandler : IRequestHandler<GetSubtitlePathById, E
         CancellationToken cancellationToken)
     {
         await using TvContext dbContext = await _dbContextFactory.CreateDbContextAsync(cancellationToken);
-        Option<string> maybeSubtitlePath = await dbContext.Subtitles
-            .SelectOneAsync(s => s.Id, s => s.Id == request.Id)
-            .MapT(s => s.Path);
+        Option<Subtitle> maybeSubtitle = await dbContext.Subtitles
+            .SelectOneAsync(s => s.Id, s => s.Id == request.Id);
+
+        foreach (string plexUrl in await GetPlexUrl(request, dbContext, maybeSubtitle))
+        {
+            return plexUrl;
+        }
+
+        foreach (string jellyfinUrl in await GetJellyfinUrl(request, dbContext, maybeSubtitle))
+        {
+            return jellyfinUrl;
+        }
         
-        // TODO: check for plex path
+        // TODO: check for emby path
+
+        return maybeSubtitle
+            .Map(s => s.Path)
+            .ToEither(BaseError.New($"Unable to locate subtitle with id {request.Id}"));
+    }
+
+    private static async Task<Option<string>> GetPlexUrl(
+        GetSubtitlePathById request,
+        TvContext dbContext,
+        Option<Subtitle> maybeSubtitle)
+    {
+        // check for plex episode
         Option<int> maybePlexId = await dbContext.Connection.QuerySingleOrDefaultAsync<int?>(
                 @"select PMS.Id from PlexMediaSource PMS
                      inner join Library L on PMS.Id = L.MediaSourceId
@@ -36,6 +59,7 @@ public class GetSubtitlePathByIdHandler : IRequestHandler<GetSubtitlePathById, E
                 new { SubtitleId = request.Id })
             .Map(Optional);
 
+        // check for plex movie
         if (maybePlexId.IsNone)
         {
             maybePlexId = await dbContext.Connection.QuerySingleOrDefaultAsync<int?>(
@@ -52,15 +76,53 @@ public class GetSubtitlePathByIdHandler : IRequestHandler<GetSubtitlePathById, E
 
         foreach (int plexMediaSourceId in maybePlexId)
         {
-            foreach (string subtitlePath in maybeSubtitlePath)
+            foreach (string subtitlePath in maybeSubtitle.Map(s => s.Path))
             {
                 return $"http://localhost:{Settings.ListenPort}/media/plex/{plexMediaSourceId}/{subtitlePath}";
             }
         }
+
+        return Option<string>.None;
+    }
+
+    private static async Task<Option<string>> GetJellyfinUrl(
+        GetSubtitlePathById request,
+        TvContext dbContext,
+        Option<Subtitle> maybeSubtitle)
+    {
+        // check for jellyfin episode
+        Option<string> maybeJellyfinId = await dbContext.Connection.QuerySingleOrDefaultAsync<string>(
+                @"select JE.ItemId from JellyfinEpisode JE
+                     inner join EpisodeMetadata EM on EM.EpisodeId = JE.Id
+                     inner join Subtitle S on EM.Id = S.EpisodeMetadataId
+                     where S.Id = @SubtitleId",
+                new { SubtitleId = request.Id })
+            .Map(Optional);
+
+        // check for jellyfin movie
+        if (maybeJellyfinId.IsNone)
+        {
+            maybeJellyfinId = await dbContext.Connection.QuerySingleOrDefaultAsync<string>(
+                    @"select JM.ItemId from JellyfinMovie JM
+                     inner join MovieMetadata MM on MM.MovieId = JM.Id
+                     inner join Subtitle S on MM.Id = S.MovieMetadataId
+                     where S.Id = @SubtitleId",
+                    new { SubtitleId = request.Id })
+                .Map(Optional);
+        }
+
+        foreach (string jellyfinItemId in maybeJellyfinId)
+        {
+            foreach (Subtitle subtitle in maybeSubtitle)
+            {
+                int index = subtitle.StreamIndex - JellyfinStream.ExternalStreamOffset;
+                string extension = Subtitle.ExtensionForCodec(subtitle.Codec);
+                var subtitlePath =
+                    $"Videos/{jellyfinItemId}/{jellyfinItemId}/Subtitles/{index}/{index}/Stream.{extension}";
+                return $"http://localhost:{Settings.ListenPort}/media/jellyfin/{subtitlePath}";
+            }
+        }
         
-        // TODO: check for jellyfin path
-        // TODO: check for emby path
-        
-        return maybeSubtitlePath.ToEither(BaseError.New($"Unable to locate subtitle with id {request.Id}"));
+        return Option<string>.None;
     }
 }