+++ ErsatzTV.Core/Jellyfin/JellyfinStream.cs
@@ -1,0 +1,6 @@
+namespace ErsatzTV.Core.Jellyfin;
+
+public static class JellyfinStream
+{
+    public static readonly int ExternalStreamOffset = 100_000;
+}