+++ ErsatzTV.Infrastructure/Data/Repositories/EmbyTelevisionRepository.cs
@@ -729,8 +729,6 @@ public class EmbyTelevisionRepository : IEmbyTelevisionRepository
             metadata.Artwork.Remove(artworkToRemove);
         }
         
-        await MetadataRepository.UpdateSubtitles(dbContext, metadata, incomingMetadata.Subtitles);
-
         // version
         MediaVersion version = existing.MediaVersions.Head();
         MediaVersion incomingVersion = incoming.MediaVersions.Head();