+++ ErsatzTV.Infrastructure/Data/Repositories/JellyfinMovieRepository.cs
@@ -345,8 +345,6 @@ public class JellyfinMovieRepository : IJellyfinMovieRepository
             fanArt.DateUpdated = incomingFanArt.DateUpdated;
         }
 
-        await MetadataRepository.UpdateSubtitles(dbContext, metadata, incomingMetadata.Subtitles);
-
         // version
         MediaVersion version = existing.MediaVersions.Head();
         MediaVersion incomingVersion = incoming.MediaVersions.Head();