+++ ErsatzTV.Infrastructure/Data/Repositories/JellyfinTelevisionRepository.cs
@@ -732,8 +732,6 @@ public class JellyfinTelevisionRepository : IJellyfinTelevisionRepository
             metadata.Artwork.Remove(artworkToRemove);
         }
         
-        await MetadataRepository.UpdateSubtitles(dbContext, metadata, incomingMetadata.Subtitles);
-
         // version
         MediaVersion version = existing.MediaVersions.Head();
         MediaVersion incomingVersion = incoming.MediaVersions.Head();