+++ ErsatzTV.Scanner/Core/Metadata/MediaServerTelevisionLibraryScanner.cs
@@ -2,6 +2,7 @@
 using ErsatzTV.Core.Domain;
 using ErsatzTV.Core.Domain.MediaServer;
 using ErsatzTV.Core.Errors;
+using ErsatzTV.Core.Extensions;
 using ErsatzTV.Core.Interfaces.Metadata;
 using ErsatzTV.Core.Interfaces.Repositories;
 using ErsatzTV.Core.MediaSources;
@@ -472,7 +473,8 @@ public abstract class MediaServerTelevisionLibraryScanner<TConnectionParameters,
                             existing,
                             incoming,
                             deepScan,
-                            None));
+                            None))
+                    .BindT(UpdateSubtitles);
             }
 
             if (maybeEpisode.IsLeft)
@@ -800,5 +802,33 @@ public abstract class MediaServerTelevisionLibraryScanner<TConnectionParameters,
         }
 
         return result;
+    }
+    
+    private async Task<Either<BaseError, MediaItemScanResult<TEpisode>>> UpdateSubtitles(
+        MediaItemScanResult<TEpisode> existing)
+    {
+        try
+        {
+            MediaVersion version = existing.Item.GetHeadVersion();
+            Option<EpisodeMetadata> maybeMetadata = existing.Item.EpisodeMetadata.HeadOrNone();
+            foreach (EpisodeMetadata metadata in maybeMetadata)
+            {
+                List<Subtitle> subtitles = version.Streams
+                    .Filter(s => s.MediaStreamKind is MediaStreamKind.Subtitle or MediaStreamKind.ExternalSubtitle)
+                    .Map(Subtitle.FromMediaStream)
+                    .ToList();
+
+                if (await _metadataRepository.UpdateSubtitles(metadata, subtitles))
+                {
+                    return existing;
+                }
+            }
+
+            return BaseError.New("Failed to update media server subtitles");
+        }
+        catch (Exception ex)
+        {
+            return BaseError.New(ex.ToString());
+        }
     }
 }